+++ Backend/Models/Company.js
@@ -1,0 +1,16 @@
+const mongoose = require("mongoose");
+const {Schema}=mongoose;
+
+const companyschema = new Schema({
+    name: String,
+    email: String,
+    number: Number,  
+    ceo: String,
+    hr: String,
+    jobsposted:[String],
+    address: String,
+    password: String,
+  });
+
+  const Company = mongoose.model("company", companyschema);
+  modules.exports=Company;
