+++ Backend/Models/Tpo.js
@@ -1,0 +1,8 @@
+const mongoose = require("mongoose");
+const {Schema}=mongoose;
+const tpoSchema = new Schema({
+    email: String,
+    password: String,
+  }) ;
+  Tpo=mongoose.model("Tpo",tpoSchema);
+  module.exports=Tpo;
