+++ package.json
@@ -39,7 +39,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "client": "npm start ",
-    "dev": "concurrently \"nodemon app.js\" \"npm run client\""
+    "dev": "concurrently \"nodemon ./Backend/app.js\" \"npm run client\""
   },
   "eslintConfig": {
     "extends": [
